Refactored pdf mocking
