+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-ws/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/ws/controller/TransformationController.java
@@ -17,9 +17,6 @@ import ch.admin.bag.covidcertificate.backend.transformation.ws.util.MockHelper;
 import ch.ubique.openapi.docannotations.Documentation;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Base64;
 import javax.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -82,11 +79,6 @@ public class TransformationController {
     @PostMapping(path = "/pdf")
     public @ResponseBody ResponseEntity<PdfPayload> getPdf(
             @Valid @RequestBody HCertPayload hCertPayload) throws IOException {
-        final byte[] pdfBytes =
-                Files.readAllBytes(Paths.get("src/main/resources/cert-pdf-mock.pdf"));
-        final var pdfEncoded = Base64.getEncoder().encodeToString(pdfBytes);
-        final var pdfPayload = new PdfPayload();
-        pdfPayload.setPdf(pdfEncoded);
-        return ResponseEntity.ok(pdfPayload);
+        return ResponseEntity.ok(mockHelper.getCertPdfMock(hCertPayload));
     }
 }
