+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-ws/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/ws/util/MockHelper.java
@@ -2,6 +2,7 @@ package ch.admin.bag.covidcertificate.backend.transformation.ws.util;
 
 import ch.admin.bag.covidcertificate.backend.transformation.model.CertLightPayload;
 import ch.admin.bag.covidcertificate.backend.transformation.model.HCertPayload;
+import ch.admin.bag.covidcertificate.backend.transformation.model.PdfPayload;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.net.URI;
@@ -11,6 +12,9 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpRequest.BodyPublishers;
 import java.net.http.HttpResponse;
 import java.net.http.HttpResponse.BodyHandlers;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Base64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -38,5 +42,14 @@ public class MockHelper {
                         .build();
         final HttpResponse<String> response = httpClient.send(request, BodyHandlers.ofString());
         return objectMapper.readValue(response.body(), CertLightPayload.class);
+    }
+
+    public PdfPayload getCertPdfMock(HCertPayload hCertPayload) throws IOException {
+        final byte[] pdfBytes =
+            Files.readAllBytes(Paths.get("src/main/resources/cert-pdf-mock.pdf"));
+        final var pdfEncoded = Base64.getEncoder().encodeToString(pdfBytes);
+        final var pdfPayload = new PdfPayload();
+        pdfPayload.setPdf(pdfEncoded);
+        return pdfPayload;
     }
 }
